+++ src/gort/recipes/operations.py
@@ -84,6 +84,7 @@ class StartupRecipe(BaseRecipe):
         self.gort.log.info("Turning off all calibration lamps and dome lights.")
         await self.gort.nps.calib.all_off()
         await self.gort.enclosure.lights.dome_all_off()
+        await self.gort.enclosure.lights.spectrograph_room.off()
 
         self.gort.log.info("Reconnecting AG cameras.")
         await self.gort.ags.reconnect()