+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next version
 
+### 🚀 New
+
+* Replace all instances of Pandas with Polars.
+
 ### 🔧 Fixed
 
 * Manually refresh the progress bar only once a second, and only after an update. Hopefully this helps with the event loop getting sluggish after a while.