+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Next version
+
+### 🚀 New
+
+* Replace all instances of Pandas with Polars.
+
+### 🔧 Fixed
+
+* Manually refresh the progress bar only once a second, and only after an update. Hopefully this helps with the event loop getting sluggish after a while.
+
+
 ## 0.11.0 - March 29, 2024
 
 ### 🚀 New