+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### 🔧 Fixed
 
 * Manually refresh the progress bar only once a second, and only after an update. Hopefully this helps with the event loop getting sluggish after a while.
+* Turn off spectrograph room lights on startup.
 
 
 ## 0.11.0 - March 29, 2024