+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 * Manually refresh the progress bar only once a second, and only after an update. Hopefully this helps with the event loop getting sluggish after a while.
 * Turn off spectrograph room lights on startup.
 
+### ⚙️ Engineering
+
+* Format code using `ruff` and updated dependencies and workflows.
+
 
 ## 0.11.0 - March 29, 2024